# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DIMO Vehicle-Signal-Decoding",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/device-config/eth-addr/{ethAddr}/jobs": {
            "get": {
                "description": "Retrieve the jobs based on device's Ethereum Address.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle-signal-decoding"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethereum Address",
                        "name": "ethAddr",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved jobs",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.JobResponse"
                        }
                    },
                    "400": {
                        "description": "incorrect eth addr format"
                    }
                }
            }
        },
        "/device-config/eth-addr/{ethAddr}/jobs/pending": {
            "get": {
                "description": "Retrieve the jobs based on device's Ethereum Address.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle-signal-decoding"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethereum Address",
                        "name": "ethAddr",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved jobs",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.JobResponse"
                        }
                    },
                    "400": {
                        "description": "incorrect eth addr format"
                    }
                }
            }
        },
        "/device-config/eth-addr/{ethAddr}/jobs/{jobId}/{status}": {
            "patch": {
                "description": "Path job status based on device's Ethereum Address.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle-signal-decoding"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethereum Address",
                        "name": "ethAddr",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Status to set",
                        "name": "status",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "incorrect eth addr format"
                    },
                    "404": {
                        "description": "Not Found - No job found with id"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_controllers.JobResponse": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Flow to get your device config: start by Getting the URL's to your device's configuration: `/device-config/eth-addr/:ethAddr/urls` or by vin, check readme for details.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SignatureAuth": {
            "type": "apiKey",
            "name": "Signature",
            "in": "header"
        }
    }
}